backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-filter==23.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
drf-spectacular==0.27.1
channels==4.0.0
celery==5.3.6
redis==5.0.1
Pillow==10.2.0
gunicorn==21.2.0
